New changes: i moved all work with courier to a separate file
