+++ game/scripts/vscripts/common/init.lua
@@ -3,6 +3,7 @@ require("common/utils")
 require("common/match_events")
 require("common/webapi")
 require("common/patreons")
+require("common/courier")
 
 require("common/disable_help")
 require("common/smart_random")
